+++ lib/rdf/dataset.ex
@@ -884,9 +884,16 @@ defmodule RDF.Dataset do
     def member?(dataset, statement), do: {:ok, Dataset.include?(dataset, statement)}
     def count(dataset), do: {:ok, Dataset.statement_count(dataset)}
 
-    def slice(dataset) do
-      size = Dataset.statement_count(dataset)
-      {:ok, size, &Enumerable.List.slice(Dataset.statements(dataset), &1, &2, size)}
+    if Version.match?(System.version(), ">= 1.14.0") do
+      def slice(dataset) do
+        size = Dataset.statement_count(dataset)
+        {:ok, size, &Dataset.statements/1}
+      end
+    else
+      def slice(dataset) do
+        size = Dataset.statement_count(dataset)
+        {:ok, size, &Enumerable.List.slice(Dataset.statements(dataset), &1, &2, size)}
+      end
     end
 
     def reduce(dataset, acc, fun) do