+++ lib/rdf/prefix_map.ex
@@ -421,12 +421,35 @@ defmodule RDF.PrefixMap do
     end
   end
 
+  @doc """
+  Converts prefix map to a list.
+
+  Each prefix-namespace pair in the prefix map is converted to a two-element tuple
+  `{prefix, namespace_iri}` in the resulting list.
+
+  ## Examples
+
+      iex> RDF.PrefixMap.new(ex: "http://example.com/") |> RDF.PrefixMap.to_list()
+      [ex: ~I<http://example.com/>]
+
+  """
+  @spec to_list(t()) :: [{prefix(), namespace()}]
+  def to_list(%__MODULE__{map: map}), do: Map.to_list(map)
+
   defimpl Enumerable do
     def reduce(%RDF.PrefixMap{map: map}, acc, fun), do: Enumerable.reduce(map, acc, fun)
 
     def member?(%RDF.PrefixMap{map: map}, mapping), do: Enumerable.member?(map, mapping)
     def count(%RDF.PrefixMap{map: map}), do: Enumerable.count(map)
-    def slice(%RDF.PrefixMap{map: map}), do: Enumerable.slice(map)
+
+    if Version.match?(System.version(), ">= 1.14.0") do
+      def slice(%RDF.PrefixMap{map: map}) do
+        size = map_size(map)
+        {:ok, size, &RDF.PrefixMap.to_list/1}
+      end
+    else
+      def slice(%RDF.PrefixMap{map: map}), do: Enumerable.slice(map)
+    end
   end
 
   defimpl Inspect do