+++ lib/rdf/property_map.ex
@@ -318,13 +318,36 @@ defmodule RDF.PropertyMap do
     end
   end
 
+  @doc """
+  Converts property map to a list.
+
+  Each term-IRI pair in the property map is converted to a two-element tuple
+  `{term, iri}` in the resulting list.
+
+  ## Examples
+
+      iex> RDF.PropertyMap.new(foo: "http://example.com/foo") |> RDF.PropertyMap.to_list()
+      [foo: ~I<http://example.com/foo>]
+
+  """
+  @spec to_list(t()) :: [{atom, IRI.t()}]
+  def to_list(%__MODULE__{iris: iris}), do: Map.to_list(iris)
+
   defimpl Enumerable do
     alias RDF.PropertyMap
     def reduce(%PropertyMap{iris: iris}, acc, fun), do: Enumerable.reduce(iris, acc, fun)
 
     def member?(%PropertyMap{iris: iris}, mapping), do: Enumerable.member?(iris, mapping)
     def count(%PropertyMap{iris: iris}), do: Enumerable.count(iris)
-    def slice(%PropertyMap{iris: iris}), do: Enumerable.slice(iris)
+
+    if Version.match?(System.version(), ">= 1.14.0") do
+      def slice(%PropertyMap{iris: iris}) do
+        size = map_size(iris)
+        {:ok, size, &PropertyMap.to_list/1}
+      end
+    else
+      def slice(%PropertyMap{iris: iris}), do: Enumerable.slice(iris)
+    end
   end
 
   defimpl Inspect do