+++ test/unit/literal_test.exs
@@ -7,7 +7,11 @@ defmodule RDF.LiteralTest do
   alias RDF.Literal.{Generic, Datatype}
   alias Decimal, as: D
 
-  doctest RDF.Literal
+  # Elixir 1.14 changed the order of inspect strings of structs, so,
+  # the literals in error strings
+  if Version.match?(System.version(), ">= 1.14.0") do
+    doctest RDF.Literal
+  end
 
   alias RDF.NS